+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `is_holiday` function
 ### Changed
 - Dependencies structure modified
 - Test system modified
 - Jalali events updated
+- `README.md` updated
 ## [0.4] - 2026-04-29
 ### Added
 - Bahesab calendar events